chat-header component layout fixes
