+++ conversion/minimal-chat/src/assets/main.css
@@ -3,21 +3,19 @@
 a,
 .green {
   text-decoration: none;
-  color: hsla(160, 100%, 37%, 1);
+  color: #00bd7e;
   transition: 0.4s;
   padding: 3px;
 }
 
 @media (hover: hover) {
   a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
+    background-color: #00bd7e33;
   }
 }
 
-
 body {
   display: flex;
-  place-items: center;
 }
 
 #app {
