+++ conversion/minimal-chat/src/components/chat-header.vue
@@ -1,0 +1,213 @@
+<!-- eslint-disable no-unused-vars -->
+
+<script setup>
+import { ref, computed } from 'vue';
+const isProcessing = ref(false); // Example processing state
+
+// Define props
+const props = defineProps({
+    selectedModel: {
+        type: String,
+        required: true
+    }
+});
+
+let storedConversations = ref(JSON.parse(localStorage.getItem("gpt-conversations")) || []);
+let isSidebarOpen = ref(false);
+let showConversationOptions = ref(false);
+
+const faConversationsCountClass = computed(() => {
+    const length = storedConversations.value.length;
+    return `fa-${length > 0 ? length : '0'}`;
+});
+
+function toggleSidebar() {
+    isSidebarOpen.value = !isSidebarOpen.value;
+}
+
+function deleteCurrentConversation() {
+    // Implement conversation deletion logic
+}
+
+function clearMessages() {
+    // Implement message clearing logic
+}
+
+function onShowConversationsClick() {
+    // Implement showing conversations logic
+}
+</script>
+
+<template>
+    <div class="header box">
+        <a v-show="props.selectedModel.includes('bison')" href="https://github.com/fingerthief/minimal-gpt#try-minimalgpt"
+            target="_blank" class="no-style-link">
+            MinimalPaLM
+        </a>
+        <a v-show="props.selectedModel.includes('claude')" href="https://github.com/fingerthief/minimal-gpt#try-minimalgpt"
+            target="_blank" class="no-style-link">
+            MinimalClaude
+        </a>
+        <a v-show="props.selectedModel.includes('gpt')" href="https://github.com/fingerthief/minimal-gpt#try-minimalgpt"
+            target="_blank" class="no-style-link">
+            MinimalGPT
+        </a>
+        <a v-show="props.selectedModel.includes('lmstudio')" href="https://github.com/fingerthief/minimal-gpt#try-minimalgpt"
+            target="_blank" class="no-style-link">
+            MinimalLocal
+        </a>
+        <a href="https://github.com/fingerthief/minimal-gpt#try-minimalgpt" target="_blank" class="no-style-link">
+            <span class="fa-brands fa-github"></span>
+        </a>
+        <div class="hover-increase-size settings-btn fa-solid fa-sliders fa-lg" @click="toggleSidebar"></div>
+        <div class="hover-increase-size trash-btn fa-solid fa-trash-can fa-lg" @click="deleteCurrentConversation"></div>
+        <div class="conversations-count">
+            <div class="general-processing" v-show="isProcessing">
+                <span class="loading spinner"></span>
+            </div>
+            <span class="hover-increase-size save-icon fa-solid fa-floppy-disk fa-xl" @click="clearMessages"></span>
+            <span class="count-icon fa-xl fa-solid" :class="faConversationsCountClass"></span>
+            <div class="hover-increase-size saved-conversations-dropdown fa-solid fa-message fa-xl"
+                @click="onShowConversationsClick"></div>
+        </div>
+    </div>
+</template>
+
+<style lang="scss" scoped>
+@use "@fortawesome/fontawesome-free/scss/fontawesome.scss";
+@use "@fortawesome/fontawesome-free/scss/solid.scss";
+@use "@fortawesome/fontawesome-free/scss/regular.scss";
+@use "@fortawesome/fontawesome-free/scss/brands.scss";
+
+/* Add your component-specific styles here */
+$icon-color: rgb(187, 187, 187);
+$shadow-color: #252629;
+$shadow-offset-x: 0px;
+$shadow-offset-y: 1px;
+$shadow-blur-radius: 2px;
+$shadow-spread-radius: 0px;
+
+
+.saved-conversations {
+    display: none;
+}
+
+.saved-conversations--visible {
+    display: flex;
+
+}
+
+.box {
+    box-shadow: $shadow-offset-x $shadow-offset-y $shadow-blur-radius $shadow-spread-radius $shadow-color;
+}
+
+.header {
+    background-color: #202124;
+    padding: 10px;
+    font-size: 18px;
+    font-weight: bold;
+    text-align: center;
+    border-bottom: 1px solid #2b2c30;
+    position: relative; // Add this line
+    border-radius: 4px;
+}
+
+
+.conversations-count {
+    display: contents;
+    position: relative;
+    left: 0px;
+
+    .count-icon {
+        padding: 2px;
+        top: 8px;
+        right: 11px;
+        color: $icon-color;
+        float: right;
+        position: relative;
+    }
+
+    .save-icon {
+        padding: 2px;
+        right: 69px;
+        top: 8px;
+        color: $icon-color;
+        float: right;
+        position: relative;
+    }
+}
+
+
+.settings-btn {
+    position: absolute;
+    left: 10px;
+    top: 50%;
+    transform: translateY(-50%);
+    background-color: transparent;
+    border: none;
+    color: $icon-color;
+    cursor: pointer;
+    outline: none;
+}
+
+.saved-conversations-dropdown {
+    position: absolute;
+    top: 50%;
+    right: 10px;
+    color: $icon-color;
+    transform: translateY(-50%);
+}
+
+.saved-conversations-dropdown select {
+    padding: 6px 12px;
+    font-size: 14px;
+    border: 1px solid #5a5a5a;
+    border-radius: 4px;
+    background-color: #3a3a3a;
+    color: #f1f1f1;
+    /* Change the text color to a lighter shade */
+    cursor: pointer;
+}
+
+.saved-conversations-dropdown select:focus {
+    outline: none;
+    /* Remove the default focus outline */
+    box-shadow: 0 0 0 1px #5a5a5a;
+}
+
+.trash-btn {
+    position: absolute;
+    left: 50px;
+    top: 50%;
+    transform: translateY(-50%);
+    color: $icon-color;
+    cursor: pointer;
+}
+
+.clearfix::after {
+    content: "";
+    clear: both;
+    display: table;
+}
+
+.clear-button {
+    width: 72%;
+    margin-bottom: 10px;
+    padding: 5px 10px;
+}
+
+
+.no-style-link {
+    text-decoration: none;
+    color: inherit;
+}
+
+.hover-increase-size {
+    cursor: pointer;
+    transition: transform 0.3s;
+}
+
+.hover-increase-size:hover {
+    transform: scale(1.1);
+}
+</style>
